+++ api/feeder-api/src/main/java/cz/inqool/dl4dh/feeder/scheduler/Scheduler.java
@@ -3,6 +3,7 @@ package cz.inqool.dl4dh.feeder.scheduler;
 import cz.inqool.dl4dh.feeder.service.ImportService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -21,5 +22,11 @@ public class Scheduler {
         log.debug("SCHEDULER STARTS");
         importService.sync();
         log.debug("SCHEDULER ENDS");
+    }
+
+    @CacheEvict(value = "collections", allEntries = true)
+    @Scheduled(fixedDelay = 600000)
+    public void emptyCollections() {
+        log.debug("Emptying collections cache");
     }
 }