+++ api/feeder-api/src/main/java/cz/inqool/dl4dh/feeder/service/CollectionService.java
@@ -1,0 +1,11 @@
+package cz.inqool.dl4dh.feeder.service;
+
+import cz.inqool.dl4dh.feeder.kramerius.dto.CollectionDto;
+
+import java.util.List;
+
+public interface CollectionService {
+
+    List<CollectionDto> getCollections();
+
+}